requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
django-environ==0.11.2
Faker==20.1.0
django-filter==23.4
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1